requirements.txt
streamlit
astrapy
python-dotenv